Working on NaSch problem for break probability = 0.9

+++ nasch/periodic_boundary_conditions/cc_nasch.h
@@ -17,7 +17,7 @@ class NaSch
  // ----------------------------------------------------------------
  // Constructor
  // ----------------------------------------------------------------
- NaSch(unsigned lane_size, unsigned maximum_velocity, double break_probability);
+ NaSch(unsigned long lane_size, unsigned maximum_velocity, double break_probability);
  
  // ----------------------------------------------------------------
  // Destructor
@@ -27,7 +27,7 @@ class NaSch
  // ----------------------------------------------------------------
  // Initialise lane configuration
  // ----------------------------------------------------------------
- void initialise(unsigned lane_size, unsigned maximum_velocity, double break_probability);
+ void initialise(unsigned long lane_size, unsigned maximum_velocity, double break_probability);
  
  // ----------------------------------------------------------------
  // Clear data structures
@@ -42,7 +42,7 @@ class NaSch
  // ----------------------------------------------------------------
  // Update vehicles list
  // ----------------------------------------------------------------
- unsigned update_vehicles_list();
+ unsigned long update_vehicles_list();
  
  // ----------------------------------------------------------------
  // Update lane based on NaSch rules
@@ -59,22 +59,22 @@ class NaSch
  // ---------------------------------------------------------------- 
  void print(bool print_velocities = false);
  
- inline unsigned lane_size() {return Lane_size;}
+ inline unsigned long lane_size() {return Lane_size;}
  inline unsigned maximum_velocity() {return Maximum_velocity;}
  inline double &density() {return Density;}
  inline double density() const {return Density;}
- inline unsigned &number_of_vehicles() {return Number_of_vehicles;}
- inline unsigned number_of_vehicles() const {return Number_of_vehicles;}
+ inline unsigned long &number_of_vehicles() {return Number_of_vehicles;}
+ inline unsigned long number_of_vehicles() const {return Number_of_vehicles;}
  
  protected:
  
- unsigned Lane_size;
+ unsigned long Lane_size;
  unsigned Maximum_velocity;
  double Break_probability;
  
  // Density associated with the number of vehicles in the lane
  double Density;
- unsigned Number_of_vehicles;
+ unsigned long Number_of_vehicles;
  
  std::vector<Vehicle*> Lane;
  std::vector<Vehicle*> Vehicles_pt;
